requirements.txt
pandas==1.2.3
numpy==1.21.2
matplotlib==3.3.4
streamlit==0.86.0
Pillow==8.3.2
scikit_learn==0.24.2
